requirements.txt
flask
google-generativeai
firebase-admin
